+++ lib/rerune.dart
@@ -1,8 +1,6 @@
 export 'src/cache/cache_store.dart';
-export 'src/controller/ota_localization_controller.dart';
 export 'src/model/manifest.dart';
 export 'src/model/ota_error.dart';
 export 'src/model/rerune_text_update_event.dart';
 export 'src/model/update_result.dart';
 export 'src/policy/update_policy.dart';
-export 'src/widget/ota_typed_localization_builder.dart';